new hash with optimistic locking scheme

+++ core/src/main/java/org/cache2k/core/WiredCache.java
@@ -339,9 +339,9 @@ public class WiredCache<K, V> extends AbstractCache<K, V>
   public int getTotalEntryCount() {
     synchronized (lockObject()) {
       if (storage != null) {
-        return storage.getTotalEntryCount();
+        return (int) storage.getTotalEntryCount();
       }
-      return heapCache.getLocalSize();
+      return (int) heapCache.getLocalSize();
     }
   }
